requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
mariadb==1.1.12
MarkupSafe==3.0.2
packaging==24.2
Werkzeug==3.1.3
gunicorn==23.0.0
boto3==1.35.99
psycopg2-binary>=2.9.0,<3.0.0
